backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
aiosmtplib==3.0.1
python-multipart==0.0.6
pillow==10.1.0
jinja2==3.1.2
pymongo==4.6.0
motor==3.3.2
